WI-068: rename prelude sort Int -> Int64
Make machine-boundedness explicit in the name. The prelude integer sort
`Int` is removed entirely (no alias) and replaced by `Int64` across
stdlib, the rust impl, codegen, tests, examples, scaland, and docs.

- stdlib: int.anthill -> int64.anthill (sort + all signatures);
  anthill-stl / cpp-gen bindings; reflect / realization / prelude refs.
- rust: PRELUDE_SORTS, integer-literal typing, builtin op names + error
  labels, cpp-gen/smt-gen op + primitive-type maps, codegen Int64->i64,
  embedded-stdlib include paths.
- scaland: Prelude/Loader primitive lists, TypeGen map, EmbeddedStdlib
  path, smt-gen op map, load-test QN-resolution refs.
- docs: spec, proposals, design notes, forward-mapping references.

Kept as `Int` (different sorts): Z3/SMT-LIB `Int`, Scala-native `Int`,
term-repr `Const(Int(n))`, and Rust `Value::Int`/`Literal::Int`.

Error[Overflow] on Int64 arithmetic is deferred to the effect-discharge
work (WI-066): the typer hard-errors on undeclared effects and no
discharge mechanism exists yet. The constant-dependent-type alternative
(Int[bits=N]) is also deferred (deep typing change).

Full rust suite green (1541 passed, 0 failed) on top of current main.

+++ rustland/anthill-cli/tests/prove_cache_test.rs
@@ -25,7 +25,7 @@ fn write_temp(name: &str, contents: &str) -> PathBuf {
 const SRC_BASE: &str = r#"
     namespace test.cache.simple
       export simple_unsat
-      entity Cfg(scale: Int)
+      entity Cfg(scale: Int64)
       fact Cfg(scale: 5)
 
       rule simple_unsat(?marker)
