+++ rustland/anthill-cli/tests/prove_outcome_test.rs
@@ -33,7 +33,7 @@ fn sat_proof_with_model_flag_emits_get_model_in_smt() {
     let src = r#"
         namespace test.outcome.sat
           export sat_witness
-          entity Cfg(scale: Int)
+          entity Cfg(scale: Int64)
           fact Cfg(scale: 5)
           rule sat_witness(?marker)
             :- Cfg(scale: ?s), gt(?s, 0), eq(?marker, ?s)
@@ -57,7 +57,7 @@ fn unsat_proof_with_cores_flag_emits_get_unsat_core() {
     let src = r#"
         namespace test.outcome.cores
           export cores_witness
-          entity Cfg(scale: Int)
+          entity Cfg(scale: Int64)
           fact Cfg(scale: 5)
           rule cores_witness(?marker)
             :- Cfg(scale: ?s), gt(?s, 99), eq(?marker, ?s)
@@ -81,7 +81,7 @@ fn no_outcome_flags_keeps_legacy_smt() {
     let src = r#"
         namespace test.outcome.legacy
           export legacy
-          entity Cfg(scale: Int)
+          entity Cfg(scale: Int64)
           fact Cfg(scale: 5)
           rule legacy(?marker)
             :- Cfg(scale: ?s), gt(?s, 99), eq(?marker, ?s)
@@ -107,7 +107,7 @@ fn sat_verdict_with_model_populates_cli_output() {
     let src = r#"
         namespace test.outcome.live
           export sat_live
-          entity Cfg(scale: Int)
+          entity Cfg(scale: Int64)
           fact Cfg(scale: 5)
           rule sat_live(?marker)
             :- Cfg(scale: ?s), gt(?s, 0), eq(?marker, ?s)
@@ -133,7 +133,7 @@ fn cache_entry_carries_model_text() {
     let src = r#"
         namespace test.outcome.cache
           export sat_for_cache
-          entity Cfg(scale: Int)
+          entity Cfg(scale: Int64)
           fact Cfg(scale: 7)
           rule sat_for_cache(?marker)
             :- Cfg(scale: ?s), gt(?s, 0), eq(?marker, ?s)
